Add cli parts for syncing, deploying, and converting to debug files
