+++ src/ModSync/ModSync.Programs/DeployToStore.cs
@@ -1,4 +1,4 @@
-namespace ModSync.ModSync.Programs;
+namespace ModSync.Programs;
 
 public class DeployToStore
 {
