--- src/ModSync/modSync.json
@@ -1,27 +1,0 @@
-{
-  "deployTargets": [
-    {
-      "name": "R2ModMan",
-      "pluginFolderPath": "C:/Users/fre/AppData/Roaming/r2modmanPlus-local/Valheim/profiles/ashlands/BepInEx/plugins/zolantris-ValheimRAFT",
-      "deployName": "R2ModMan Main",
-      "folderName": "R2ModMan"
-    },
-    {
-      "name": "ValheimServer",
-      "pluginFolderPath": "C:/steam/steamapps/common/Valheim/BepInEx/plugins",
-      "deployName": "Server Main",
-      "folderName": "Server"
-    }
-  ],
-  "runTargets": [
-    {
-      "name": "ValheimClient",
-      "binaryTarget": "C:/steam/steamapps/common/Valheim/valheim.exe",
-      "args": [
-        "-someArg",
-        "-nographics"
-      ],
-      "isConditional": true
-    }
-  ]
-}
